Changed a couple more things in index.js and moved some setup for inputs for pictures to the setup_url_input.js file.
